backend/layers/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
Deprecated==1.2.15
googleapis-common-protos==1.66.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
protobuf==5.29.3
requests==2.32.3
typing_extensions==4.12.2
wrapt==1.17.2
zipp==3.21.0
boto3==1.10.34
pyjwt==2.10.1
psutil==7.0.0
